Update onclientitemexpand.md
The OnClientItemExpand event handler does not actually fire after the panel item is fully expanded as stated in the existing text. Suggested change (edits may be needed) explains that order of events and provides an example of how one might implement an event handler that is fired after the item is expanded. A similar edit might be applied to other RadPanelBar event handler help, especially for RadPanelBar OnClientItemClicked.